+++ src/onegov.election_day/onegov/election_day/forms/__init__.py
@@ -5,6 +5,7 @@ from onegov.election_day.forms.subscribe import SubscribeForm
 from onegov.election_day.forms.upload import UploadElectionPartyResultsForm
 from onegov.election_day.forms.upload import UploadMajorzElectionForm
 from onegov.election_day.forms.upload import UploadProporzElectionForm
+from onegov.election_day.forms.upload import UploadRestForm
 from onegov.election_day.forms.upload import UploadVoteForm
 from onegov.election_day.forms.upload import UploadWabstiMajorzElectionForm
 from onegov.election_day.forms.upload import UploadWabstiProporzElectionForm
@@ -31,6 +32,7 @@ __all__ = [
     'UploadElectionPartyResultsForm',
     'UploadMajorzElectionForm',
     'UploadProporzElectionForm',
+    'UploadRestForm',
     'UploadVoteForm',
     'UploadWabstiMajorzElectionForm',
     'UploadWabstiProporzElectionForm',