+++ src/onegov.election_day/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds a REST interface to upload internal formats.
+  [msom]
+
 1.12.2 (2017-06-13)
 ~~~~~~~~~~~~~~~~~~~